Changed a casting from double to float.
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@161 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/wavelet.cpp
@@ -769,7 +769,7 @@ Wavelet::findOptimalWaveletScale(fftw_real ** synt_seis_r,
     seisNorm[i] = 0.0f;
   }
 
-  float minSeisAmp = 1e-7;
+  float minSeisAmp = static_cast<float> (1e-7);
   int totCount=0;
 
   for(int i=0;i<nWells;i++)
